requirements.txt
astroid==2.2.5
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dash==0.39.0
dash-core-components==0.44.0
dash-daq==0.1.0
dash-html-components==0.14.0
dash-renderer==0.20.0
dash-table==3.6.0
decorator==4.3.2
Flask==1.0.2
Flask-Compress==1.4.0
idna==2.8
ipython-genutils==0.2.0
isort==4.3.15
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.1
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
nbformat==4.4.0
numpy==1.16.2
plotly==3.7.0
pylint==2.3.1
pyorbital==1.5.0
pyrsistent==0.14.11
pyserial==3.4
pytz==2018.9
requests==2.21.0
retrying==1.3.3
scipy==1.2.1
six==1.12.0
traitlets==4.3.2
typed-ast==1.3.1
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1
